Initial transform data structure changes
